integrate nonfunction-type-alias experiment tests
