--- test_data/rules/experiments/nonfunction-type-aliases/analysis_options.yaml
@@ -1,3 +1,0 @@
-analyzer:
-  enable-experiment:
-    - nonfunction-type-aliases
