--- test_data/rules/experiments/nonfunction-type-aliases/rules/.mock_packages
@@ -1,1 +1,0 @@
-meta:../../../../mock_packages/meta/lib/
